[FIX] l10n_be_coda: compute check digits properly when generating IBAN from BBAN number
The check digits computation was wrong, leading to an error message when trying to import a CODA file using a BBAN number to identify the account in an IBAN bank journal. This was due to the fact the computed IBAN did not match the one contained in the journal's bank account number.

Fixes issue 1928435

closes odoo/enterprise#3468