[IMP] l10n_mx_edi: Support decimal currency based on the SAT.
The SAT provides a catalog, where indicate decimal quantity supported for
the attributes `Monto`, `ImpSaldoAnt`, `ImpPagado` and `ImpSaldoInsoluto` in
the payment complement, and the quantity is different for each currency.

0, 2, 3 or 4 are the possible values, and the more common is 2.

What is the solution provided?

Currently, the currency decimal places is used, but that value is not
related to the SAT catalog. And based on the Odoo method `precision_get()`
that if do not found a record with the name provided return `2` (that is
the common value), or the value provided for it, is used that model. If
some company require to use other currency that has other decimal quantity,
they can create a new record with the currency name and the quantity of
decimals supported, and with that, they could generate without problem
the complement.

The next error in the payment complement is fixed  when the
currency decimal places is different of 2 and the currency is `MXN` or
`USD` for example:
```Mensaje: El valor del campo Monto debe tener hasta la cantidad de
decimales que soporte la moneda registrada en el campo MonedaP.```

closes odoo/enterprise#3473