[FIX] voip: Unblock voip if connected
Before this commit, if SIP fails to connect to the websocket at the first try, the phone widget was blocked without the possibility to unblock it.
In addition, the overlay blocked all the widget, including the close icon.

Now, the phone is blocked by default, and unblock as soon as the websocket is connected.

OPW-1931725

closes odoo/enterprise#3495