[FIX] delivery_fedex: harmonized code and custom values.
The harmonized code is not passed to the commodities when internationnal
deliveries are made through fedex, this leads to wrong computed delivery
cost and create unexpected cost for users.

So we've added the harmonized code to the commodities, and also the
custom values, which is the result of unit price multiplied by quantity.

OPW-1930505

closes odoo/enterprise#3496