[FIX] voip: cancel and reject failed
This commit completes f27626607b0e4d372a3cd943940617b9cd6d9568 when the call is rejected by closing the dialog.

This commit also fix the case when a call already finished is cancelled. An error is thrown by the lib:
https://github.com/odoo/enterprise/blob/d1d79944cd3d9a76450184b42949e9bd4d2351d1/voip/static/lib/sip.js#L6863

But the error is not relevant for the end user. So the issue is catched and logged.

opw-1929582

closes odoo/enterprise#3500