[FIX] account_taxcloud: capture and refund when validating
Taxes are due based on when invoices are validated. Whether or not
they're paid doesn't matter to US states. The Captured call must
therefore happen when we validate, this is when TaxCloud considers the
invoice ready to be reported. Without this TaxCloud consistently under
reports taxes to the states.

The same is true for refunds. They should be reported as refunds as
soon as they're validated, not when they're paid.

Because this is happening in a stable release the code in
action_invoice_open remains. It will result in Odoo
capturing/refunding invoices twice, but TaxCloud will ignore the
second call. It can't easily be removed, because otherwise unpaid
invoices validated before this patch would never be captured.

closes odoo/enterprise#3514