[FIX] account_reports: make caret options work properly for financial report lines using a domain and grouping journal items by account
Before that, when defining such a financial report line, clicking on the "Journal Items" caret option for a group line displayed *all* the journal items made within the date range of the report, disregarding their journal or the fact they respect the financial report line's domain. This lead to inconsistent displayed data between the report lines and the journal items displayed for them.

closes odoo/enterprise#3511