[FIX] account_batch_payment: SEPA file generation raise AccessError
When generating a SPEA using odoo, the generated xml file is
matched as a svg file and odoo refuses to store it if the user
isn't admin.

This behavior exists to prevent the injection of forged xml files
that could be evaluated as scripts by browsers. As the SEPA file
is fully generated by odoo using lxml.etree, it is safe to bypass
the validation.

closes odoo/enterprise#3522