[FIX] stock_barcode: Scan unknown barcode on picking with reserved lines
Considering the following use-case:

On a picking with some reserved lines, if a reserved product was scanned, then a wrong barcode was scanned,
the reserved line of the product previously scanned was incremented.

This was happening because when _step_product fails, _step_lot is called but doesn't check if the product is tracked,
and thus was considering the scanned barcode as a lot/serial number.

To fix this we simply rejected the deferred if the product is not tracked in _step_lot in order to use the callback from
_step_product to return an 'Incorrect Barcode' error message

TaskID: 1882969