[IMP] l10n_mx_edi: Method that update SAT status improved
The method `l10n_mx_edi_update_sat_status()` in the invoice could be called
in any PAC status whenever the record have a CFDI related
(l10n_mx_edi_cfdi field).

For example the next case:
A CFDI calls the method to cancel, but return an error with the PAC, and
the `PAC status` is `to_cancel`, In this case I need to check if the
CFDI was cancelled in the SAT, and is necessary that method.

Also was removed the `is_required()` validation, for the next reason:

A CFDI that was send me from by supplier could be imported on a vendor
bill, and could be assigned in the field `l10n_mx_edi_cfdi`, if I want
to know the SAT status I could use the same method and return that
status.

closes odoo/enterprise#3543

closes odoo/enterprise#3967

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>