[FIX] IoT : Path to the IoT pedal is unreachable
Revision [1] broke the pedal by changing the `ip` related field
definition on the workorder: it used to return `ip_url` but it now
returns only the `ip`.

The `triggers` object used to contain url to which the javascript
would append the `'/hw_drivers/owner/check'` part. This part was
broken due to the change. We now build the URL at the start of the
widget while taking care of handling the https.

[1] 6813b89173db3318336deb8cfb3c4a6e167e9736

closes odoo/enterprise#3549