[FIX] web_studio: get the correct view arch when creating an inline view
Use studio to create an inline view.
It loads the initial view to inherit it so that modifications can be safely done.
However the method used applied some postprocess that introduced model modifiers,
as well as removing inline views.
What we want is the arch with the inheritance applied.

Steps to reproduce:
Go to a survey page, start studio to edit the view;
edit the form view of the many2on question.
When the view loads, it is not the same view that was before studio launched.
The issue is that when loading the form view for survey.question, it contains an
inline view for labels_ids and labels_ids_2.
These inline views are removed from the view by postprocess_and_fields,
and put into ['fields_views'][view_type]['fields'][field]['views'],
where viewtype is the parent view type, and field is the inline view field.

opw 1906891
opw 1933866
opw 1919788

closes odoo/enterprise#3436