[FIX] web_studio: normalize without direct named anchor
In studio, there is a normalization of studio customization done so:

- the studio customizations do not get too monstruous with eg. same
  element in the same view added and removed dozen of times
- change are minimal => to have customization done after studio that
  works as much as possible
- changes are well targeted => so future customization have less
  probability of breaking existing studio customization
- ...

In b1585e0c the code was adapted to avoid customization issue such as:

 original view: `<div><div class="original"/></div>`
 studio view:

```
<xpath expr="/div/div" position="before"><div name="before"/></xpath>
<xpath expr="/div" position="inside"><div name="after"/></xpath>
```

that resulted in a normalization:

```
<xpath expr="/div" position="inside"><div name="before"/></xpath>
<xpath expr="/div" position="inside"><div name="after"/></xpath>
```

not expected since the "before div" is now after "original div".

With b1585e0c the resulting view was:

```
<xpath expr="/div/div[1]" position="before"><div name="before"/><xpath>
<xpath expr="/div/div[1]" position="after"><div name="after"/><xpath>
```

which is wrong since the first xpath is applied first, the second adds
the "after div" after the first div which is now "before div".

The issue and previous issue stems from the fact that the computed xpath
in in the current system is expected to work for the original view, but
not for the view with all previous xpath applied which is what happen in
reality (and may change the value of the xpath we want).

With this change, we remove the system in b1585e0c and instead use
[not(@name)] when an unclear xpath has to be computed, this way the
xpath works for the original view and original view with previous xpath
adding nodes applied (since elements added in studio receive a name).

With this change, the example above results in:

```
<xpath expr="/div/div[not(@name)][1]" position="before"><div name="before"/><xpath>
<xpath expr="/div/div[not(@name)][1]" position="after"><div name="after"/><xpath>
```

which has the same result as the studio customization before
normalization.

opw-1932283
closes #3565