[FIX] l10n_mx_edi: Invoice assignation from statement was improved
When a payment is created from a bank statement, are assigned the
invoices related with the payment to allow generate the payment
complement.

But, fail in the next case:
1. Generate a customer invoice
2. Generate your payment and validate
3. Go to bank statement and generate a new record to reconcile with the
payment created in the step 2.

In this case the invoices already was added in the payment, and in the
method process_reconciliation() is not possible to get that invoices.

Now not overwrite the invoice_ids field, only extend the content if found
an invoice record.

closes odoo/enterprise#3570

closes odoo/enterprise#3989

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>