[FIX] delivery_ups: invalid dimensions
- Configure UPS US to use:
  UPS Service Type          : UPS Worldwide Express
  UPS Default Packaging Type: UPS 25KG Box
- Create a picking to be delivered to an address outside US.
- Validate the picking

An error arises from UPS:
  "Missing or invalid Combination of package length, width, and height
   values for package 1."

By default, the dimensions of the package are set to zero since they are
not specified, meaning that Length, Width and Height are set to an empty
string.

This doesn't seem accepted by UPS anymore. Actually, since we use a
pre-defined UPS package, this information is not necessary at all.

opw-1920496

closes odoo/enterprise#3578