[ADD] l10n_uk_reports_hmrc: wizard send VAT return to HMRC
We add a button to the existing tax report "Connect / Send"
to HMRC in order to send the tax information to the government.

The first time you connect, you have to log in into the hmrc site
first.  For that, we needed a proxy on our site to recuperate its
response (as they support only a fixed redirect uri).  The proxy
will check if a valid db uuid is given.  The proxy itself generates
a token and you get a token back from hmrc.  Every 4 hours, the token
needs to be refreshed.

We use two endpoints of the API: retrieving the obligations (when
and for which period vat returns (still) need to be filled in) and
submitting the vat return itself.

When you click the send button after a successfull login, a wizard
is loaded where if possible the obligation is selected which conforms
with the dates you chose in the report.  There you can send it.

For v10, we could not launch the send wizard from the form easily,
so we created a menuitem for it.  Also the options are generated through
the .context object.

opw-1819495

closes odoo/enterprise#3917

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>