[FIX] account_reports: encode filenames to follow HTTP standards
The get_report method in FinancialReportController never heard of the evils of
code duplication.
Instead of creating the response in a simple function, it simply does it 7 times,
once per file type.
Unsurprisingly, 3 times out of 7 the filename was not properly encoded.
As a result the Javascript would crash client side on an unexpected header,
if the filename contained non-US ASCII characters
(for instance if the report name was translated in German, with funny accents).

Exact same thing as commit 6a8f19c8d40b1fee0f0b0389e5aaef35b412535b,
because FinancialReportController apparently never learns.

opw 1932123

closes odoo/enterprise#3593