[FIX] account_bank_statement_import_csv: Proper compute amount from credit and debit columns
Formerly the amount of bank statements were computed as "debit - credit". This
because Odoo interprets debits as accreditations and credits as subtractions
(as in accounting). This was generating the next errors:

1. bank substraction operations were imported as positive amounts
2. bank accreditation operations were imported as negative amounts

Bank statements files are generated by banks, and banks manage debit and
credit columns different from what Odoo use it:

* credit are accreditations operations
* debit are subtractions operations

In order to import properly the bank statemens files to Odoo we update the
formula for compute the amount to "abs(credit) - abs(debit)".

NOTE: The abs() was neccesary because depending of the bank, some bank files
have the credit column expresed with negative numbers which was causing a bad
compute "credit - (-debit)".

closes odoo/enterprise#3602

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>