[IMP] l10n_mx_edi: Write the TZ time only when is required
The time on the invoice is assigned based on the journal TZ, to be
consistent with the time in that place.

But when the invoice time already exist, the time do not must be
updated, example the next case:

1. Create an invoice with a specific date
2. You wish that the time in the CFDI is '01:02:03', and assign that
value with a server action.
3. The invoice is validated, and the module is designed to do not change
the time on the invoice when the time exist. But if in this case the
journal ZIP is marked to use a different TZ, in this case, how the time
already exist the hour do not must be changed.

closes odoo/enterprise#3614