[IMP] l10n_mx_edi: Support read CFDIs from the SAT
The SAT has a tool to allow stamp the CFDIs free to some
contributors, and in the CFDIs the SAT the system generate added the
`CFDIRegistroFiscal` complement, but this has a XML error, they add the
`schemaLocation` with `xmlns:schemaLocation` and must be
`xsi:schemaLocation`.

To support this cases is fixed the XML when converting this to `etree`, to
avoid that the documents cannot be read.

Example:

```
  <cfdi:Complemento>
    <registrofiscal:CFDIRegistroFiscal xmlns:schemaLocation="http://www.sat.gob.mx/registrofiscal http://www.sat.gob.mx/sitio_internet/cfd/cfdiregistrofiscal/cfdiregistrofiscal.xsd" Version="1.0" Folio="XXX" xmlns:registrofiscal="http://www.sat.gob.mx/registrofiscal"></registrofiscal:CFDIRegistroFiscal>
    <tfd:TimbreFiscalDigital xmlns:tfd="http://www.sat.gob.mx/TimbreFiscalDigital" xsi:schemaLocation="http://www.sat.gob.umx/TimbreFiscalDigital http://www.sat.gob.mx/sitio_internet/cfd/TimbreFiscalDigital/TimbreFiscalDigitalv11.xsd" Version="1.1" UUID="XXX" FechaTimbrado="2018-01-31T14:42:21" RfcProvCertif="SAT970701NN3" SelloCFD="XXX" NoCertificadoSAT="00001000000403258748" SelloSAT="XXX" />
  </cfdi:Complemento>
```

Since the user could attach a file in the invoice and assign the same name
in the record and with this read the fiscal information from the CFDI, now
is fixed the CFDI in that cases to avoid a traceback.

closes odoo/enterprise#3615