[FIX] web_studio: enable grouping
- Install Project
- Go to the Tasks kanban viewn open Studio
- Untick 'Enable Stages'
- Tick 'Enable Stages'

The grouping is now done on the field `x_stage_id`. Several features,
such as sending messages when the state is changed, are not available
anymore.

An untold convention is that the `stage_id` field is used for grouping
in Kanban. Therefore, we use `stage_id` if it exists.

opw-1924899

closes odoo/enterprise#3622