[FIX] web_studio: avoid undeterministic bug on runbot
The issue here was in the new report tour. It requires to create a new record
to correctly display the report but Studio was opened before the record creation
so it triggers a modal with 'The record has been modified, ...' which prevented
the tour to continue.

We thus add an `extra_trigger` to wait for the record to be saved (the form to be
re-rendered in readonly mode).

Task 1937583

closes odoo/enterprise#3597