[FIX] voip: click phone number voip parameter missing
When a VOIP is configured for the company, and a user don't use VOIP
(he didn't configure username and password in the user settings).
When the user click in a phone number:

Before this commit, the pop-up of the VOIP widget is open, a yellow
message telling that the user is calling the number appears, and a
traceback is raised.

Now, the default configuration of the browser is called.

opw-1937924

closes odoo/enterprise#3640