[FIX] web_cohort: datetime localization
In the CRM cohort view, record created today are listed on the +1
column.

This is because the column start date is based on a datetime field. When
performing the read_group the datetime is localized, but it is later
truncated to keep only the date. For example, on GMT +1, the date
`2018-12-01` is converted to `2018-11-30 23:00:00`, which is then
truncated to `2018-11-30`.

By setting a `False` timezone, `tz_convert` will be `False` at:

https://github.com/odoo/odoo/blob/ef03db9edd9472201cb2c08a32d20ff0f33a5fdf/odoo/models.py#L1906

which will prevent TZ conversion.

opw-1927148

closes odoo/enterprise#3641