[FIX] web_cohort: cell domain
On the Subscription cohort view with demo data, most of the cells are at
100%. However, when clicking on a cell no record is shown.

The cell domain is not correct.

In `retention` mode, we want records with an end date larger than the
column end date. This implies that records with no end date should also
be taken into account.

On the other hand, in `churn` mode, we want the opposite domain: records
with an end date smaller than the column end date.

opw-1927148

closes odoo/enterprise#3642