[FIX] l10n_mx_reports: Avoid traceback when is opened the journal items
The parameter `financial_group_line_id` was removed when try to open the
journal items from the Trial Balance report for Mexico.

This report uses a financial lines that has a domain for the account tag,
and that parameter is not found when try to search in account.move model.

Avoid the next traceback:
`invalid literal for int() with base 10: 'MX'`

And:
```
raise ValueError("Invalid field %r in leaf %r" % (left, str(leaf)))
ValueError: Invalid field 'tag_ids' in leaf "<osv.ExtendedLeaf: ('tag_ids', 'in', [74]) on account_move_line (ctx: )>"
```

closes odoo/enterprise#3645