[FIX] account_reports: Parse correctly journal IDs
When a journal is displayed from the report view, its ID is parsed
expecting it in the form `account_<ID>`, e.g. `account_15`. However,
that will not always be the case, as it could be `level_one_<ID>`.
Causing an error when extracting the ID.

This change causes the ID to be taken from the last word, instead of the
second one, to ensure it's always parsed correctly.

closes odoo/enterprise#3646