[FIX] documents_account: company terms appear on vendor bills
When creating a vendor bill via Documents, the created bill show
the terms and conditions of the company.

The problem is due to the way terms are implemented, they set a
different default value for the `comment` field of
`account.invoice`. Value explicitely set as False when created
from the Purchase app but not set when created from Documents.

opw-1936083

closes odoo/enterprise#3651