[FIX] delivery_easypost: specify package dimension UoM
Package dimension to easypost should always be sended in inches.
However on the delivery package's form view the length/width/height
fiels do not display the UoM that the user should use.

The user could use another UoM or think that the UoM to use
is the same than on the product form view.

closes odoo/enterprise#3661