[FIX] sale_subscription_dashboard: random translation of revenue KPIs
In some cases, the dashboard of the revenue KPIs is translated in a
language not available in the database.

This is due to the fact that STAT_TYPES and FORECAST_STAT_TYPES are
translated at first variable initialization and then reused as globals.

This fix "delays" the translation by hacking _().

opw-1932209
opw-1907122

closes odoo/enterprise#3663