[FIX] l10n_mx_edi_external_trade: wrong Qty UMT
Have a product : 700ml bottle of alcoholic drink
with Unit(s) as Unit of Measure and in CFDI configuration:
 * 50202200 Bebidas alcohólicas as Code SAT
 * 22089002 Bebidas alcohólicas.... as Tariffic Fraction
 * Liter(s) as UMT Aduana
 * 0.0 as Weight

Have an Invoice for an international customer, with this product with:
 * 672 as quantity
 * 12 as price
 * 470.400 as Qty UMT (The UMT Aduana unit of measure is in litres, our
    product is in units, to transform from one to the other we need to
    multiply 672 units * 0.700 litres.)

Validate the Invoice.

Before this commit, the Qty UMT is re-writed to 470.478.

This occurs because the function _set_price_unit_umt, calculates the
unit price with the UMT unit of measure (price unit UMT) based on the
Qty UMT (Qty * price / Qty UMT). After it uses this price unit UMT to
calculate the Qty UMT. Unfortunately, as the price is rounded, the
new Qty UMT is not correct.
This second operation is redundant, the Qty UMT must be encoded by the
user or in some cases be the same as the quantity in the invoice, but
not calculated from the price unit UMT.

Now, the second operation is removed, the Qty UMT stays 470.400

opw-1932905

closes odoo/enterprise#3682

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>