[FIX] l10n_mx_edi_external_trade: wrong ValorDolares and TotalUSD
Have a product : 700ml bottle of alcoholic drink
with Unit(s) as Unit of Measure and in CFDI configuration:

* 50202200 Bebidas alcohólicas as Code SAT
* 22089002 Bebidas alcohólicas.... as Tariffic Fraction
* Liter(s) as UMT Aduana
* 0.0 as Weight

Have an Invoice for an international customer, with this product with:

* 672 as quantity
* 12 as price
* 470.400 as Qty UMT (The UMT Aduana unit of measure is in litres, our
  product is in units, to transform from one to the other we need to
  multiply 672 units * 0.700 litres.)

The SubTotal is 8064.00 MXN.

The currency of the invoice is in MXN, and the exchange rate between MXN
and USD is 1.0.

Validate the Invoice.

Before this commit, the TotalUSD and ValorDolares are 8063.99 USD.

This occurs because there is an error in the calculation. The correct
calculation method is explained in this document
(http://omawww.sat.gob.mx/informacion_fiscal/factura_electronica/Documents/Complementoscfdi/GuiaComercioExterior3_3.pdf).
We can resume it as follows :
ValorDolares : it depends of the currency  (p. 62-63):
 - if currency is MXN: ValorDolares = Importe (subtotal without discounts) / TipoCambioUSD
 - if currency is USD: ValorDolares = Importe
 - if currency is anoter: ValorDolares = Importe x TipoCambio / TipoCambioUSD
There is a common mistake to mutiply the Qty UMT with the unit price UMT. (p. 76)

TotalUSD : must be the sum of all the Valor Dolares fields (p. 48)

Now, the TotalUSD and ValorDolares are 8064.00 USD.

opw-1932905

closes odoo/enterprise#3684