[FIX] delivery_easypost: get rate with service
Usecase to reproduce:
- Create a delivery carrier easypost with a default service
- Ask for a rate request

Traceback due to [0] on a dictionary.
It happens because the code already select item [0] after
searching if a rate is availbale for current service.

Fix it by filter on the rate. Then check if rates exists for
wanted service then select the rate.

Add a test in order to include services on carrier. However Fedex
services could be modify in future. Add a warning logger if the
service do not exist anymore.

closes odoo/enterprise#3688