[FIX] sale_ebay: create new delivery addresses from transaction
Have a partner P in database with address A, and shipping address AS.
Have a sale made through ebay to a user with email address E = P.email.
Then P is reused as the partner with which the sale is done.

Before, it used the shipping address AT given in the ebay transaction data to
overwrite P's address A.
However if P already had a different shipping address set this one would be used
on the Sale Order, instead of the new one, i.e. AS instead of AT
(and A would be lost).

Now, we see if P already has a matching address, and explicitly set it
on the order as partner_shipping_id on the sale order.
If no address matches and the user already has an address, we create a new one.
In our example P would thus have address A, and child contacts with addresses
AS and AT.

opw 1930716

closes odoo/enterprise#3704