[FIX] web_studio: normalize diff -[not(@name)] ea97739
In recent change ea97739896, something was not taken into account when
in the diff lines removed are after line added that would be targetted
with eg. div[not(@name)].

The issue comes from the difference between Element.find and
Element.xpath, the former only support a subset of xpath and fails on an
expression such as [not(@name)].

This change fix the particular issue by using xpath method.

Without change, the added test failed with:

  ERROR: test_view_normalization_35
  Traceback (most recent call last):
    ...
    `   File "enterprise/web_studio/models/ir_ui_view.py", line 226, in normalize
    `     current_xpath_target = old_view_tree.find('.' + xpath.get('expr'))
    ...
    ` SyntaxError: invalid predicate

opw-1943155
closes #3699