[FIX] sale_ebay: do not fail on product with invoice_policy set to delivery
Let P be a product with invoice_policy set to 'delivery'. List P on ebay.
After P is sold, Odoo synchonises the sale to create a sale order.
After the SO creation, it creates a delivery and tries to create an invoice.
However, if all products are set to delivery, there is nothing to invoice,
so the invoice creation fails with a user error 'There is no invoiceable line'.
As a result the synchronisation of the transaction fails.

We simply do not create the invoice in such cases, since it can be created only
after the delivery is manually validated.

opw 1889746

closes odoo/enterprise#3700