[FIX] web_studio: KanbanEditor with async widgets
Before this rev., in crashed in Studio when the user tried to enter
the KanbanEditor on a kanban view with an async widget. For
instance, do F5 (to ensure that lazy-loaded scripts are cleared),
enter Studio from the Home menu, enter Accounting. The
JournalDashboardGraph widget is async because it lazyloads nvd3.

The crash occured for two reasons.

First, since rev. 5faec34a, widget's $el doesn't exist before start.
This means that for async widgets, one can't interact with the
widget's $el right after calling appentTo. This is exactly what was
done in the KanbanRecordEditor.

Second, the rendering of the KanbanEditor didn't wait at all for
async widgets to be ready. This caused a crash with the
JournalDashboardGraph because on_attach_callback was called before
the widget's $el existed.

opw-1925079
opw-1925479

Fixes odoo/odoo#30087
Closes odoo/odoo#31254

Related to odoo/odoo#31327

closes odoo/enterprise#3701