[FIX] project_timesheet_synchro: force data export in UTC
A bug was introduced since https://github.com/odoo/odoo/commit/c819fc6

Steps to reproduce:
- In desktop, create a new timesheet line for a task
- Open AS extension and try to add 1 hour
- Force the synchronisation in AS
- The timesheet line will be back to the previous value

In AS extension, the record will be updated if the write date in the
local storage is smaller than the one of the server.

Before this commit, the datetime sent by the server was in the user
timezone and was compared to the local storage write date stored in UTC.
So depending of the user timezone, the record was wrongly updated.
See update case:
time_module.str_to_datetime(ls_aal.write_date) < time_module.str_to_datetime(sv_aal[7])

Now, we force data export to UTC.

closes odoo/enterprise#3709