[FIX] sale_ebay: do not overwrite shipping name by ebay name
Commit 0f276c0e218e7593431bc65fe9c621b1e7cfb9e7 split the information taken
from an ebay transaction between partner_data and shipping_data.
So we have partner_data['name'] = transaction['Buyer']['UserID'],
the ebay pseudonym of the buyer,
and shipping_data['name']
is transaction['Buyer']['BuyerInfo']['ShippingAddress']['Name'],
the name of the person that is used for the shipping address.
Before, the latter would override the former.

Here since we split the two we overwrite the shipping address we just wrote.

opw 1930716

closes odoo/enterprise#3723