[FIX] delivery_easypost: service not available message
The system try to mix the error message and the warning message (only
one is displayed, mix them bring more precision to the user). However in
this case the warning message could be a boolean and we can't mix
boolean and string.

The merge between error and warning message is always done at the end.
So we could remove it in this part of the code since even if it works
the warning message would be duplicated.

opw-1944080

closes odoo/enterprise#3729