[FIX] l10n_be_reports: VAT report XML export: don't double amounts for grids 46 and 48
Before that, we generated the XML by grouping the report lines depending on the start of their names. Until v12, it worked fine, since the title lines for grids 46 and 48 did not have any total amount. So, to compute the total of grid 46 in the XML, we summed 46T, 46L, and 0 for the title. In 12, we added a total to the title line, without adapting the computation of the XML file. This resulted in the total amount reported in the file for both tags 46 and 48 to be doubled.

closes odoo/enterprise#3732