[FIX] sale_ebay: refactor synchronisation to avoid reaching API query limits
EBay allows to query the seller list for time ranges of at most 120 days.
This was probably wrongly interpreted as
"ebay does not allow to query the seller list for items older than 120 days".
As a result of this misinterpretation, for old listings one query was done for
each product template.
Given that by default the API query limit is 5000 per day,
it could be easily reached for clients with thousands of products.

Note that it is possible to increase the API limit by going through
https://developer.ebay.com/support/app-check
(it must be done by the client for his Dev key)

opw 1924454

closes odoo/enterprise#4147

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>