[FIX] sale_ebay: do not overwrite product template related values
When a product.template is created, it goes through:
template.with_context(create_from_tmpl=True).create_variant_ids()
which, as a result, overwrites the fields that are defined as
related='product_variant_ids.FIELD'.
(This happens even if the variants are not activated.)
To avoid that, there is a create override that takes the related values
and rewrite them after the create, in
odoo/odoo/commit/c3f81c527ddb5153f5a1eceb5e87efeb533129b0
We do the same thing here for the ebay fields.

opw 1924454

closes odoo/enterprise#3750