[FIX] web_studio: duplicate report uses old xml-id
To reproduce:

1) Edit a sale order within studio
2) Go to the reports tab
3) Duplicate a report
4) Edit the duplicated report
5) Remove it
6) Reduplicate the original report
7) The re-duplicated report is the deleted one instead of a new one

The problem is due to the generation of the `copy_no` number that is
used to generate a unique xml-id. The `copy_no` value is generated by
getting the biggest copy suffix (ie 2 for `report_saleorder_copy_2`) and
by incrementing that value. If a report is deleted, then the biggest id
is the previous one, incrementing it gives the exact same `copy_no` than
the deleted one thus load the wrong template.

The fix has been to retrieve the `copy_no` value from the copied view,
the view itself uses a `_copy_x` as suffix for its `key` field to
diffenriate it from the other views.

opw-1936088

closes odoo/enterprise#3757

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>