[FIX] documents: typo description Financial folder
The default description for Financial folder was:

```
forward bills to <strong>finance-inbox@yourcompany.com</strong> to
import them automatically.
```

this make no sense because this is the wrong domain, and the domain
does not change when the alias domain change.

With this changeset, we only say this is "inbox-financial" email alias.

opw-1946096
closes #3763

Signed-off-by: "Nicolas Lempereur (nle)" <nle@odoo.com>