[FIX] account_reports: local variable 'headers' referenced before assignment
If the user customized web.external_layout to remove
the header and footer,
trying to print an accounting report with `minimal_layout` to False
e.g. a customer statement
resulted to the exception:
```
local variable 'headers' referenced before assignment
```

This is simply because `headers` wasn't instanciated
as `root.xpath(match_klass.format('header')` returned empty
and no `lxml.etree.XMLSyntaxError` exception was raised.

This revision make sure `headers` and `footer` are
instanciated with default values before attempting to
use them.

Besides, `_run_wkhtmltopdf` requires bytes, not str,
so even in case of exception, headers was set to
a str, which then resulted to an encoding issue.

opw-1948077

closes odoo/enterprise#3783

Signed-off-by: Denis Ledoux <beledouxdenis@users.noreply.github.com>