[FIX] pos_blackbox_be: load qweb only for blackbox configs
The assets are different between a session with a blackbox install
(blackbox_pos_production_id is set on the pos.config) and for "vanilla" configs.

The xpath adds some js code that will remove 'pos_blackbox_be' from the list of
modules to load in the call to /web/webclient/qweb?mods=...

In 9.0 the load_qweb code was executed after the rendering and evaluation of the
`point_of_sale.index` page.
In 11.0, the load_qweb was executed earlier, as soon as web.asset_backend is
loaded, meaning the include is done too late.

Force the inclusion of the custom js before web.asset_backend to be sure the
blackbox qweb code is not loaded on vanilla point of sales.

This commit is the proof this implementation is very fragile.
A more future-proof version could modify pos_blackbox_be.xml to trigger the
changes only when it makes sense.
However, this approach would require to patch every instanciation of the widgets
used in the file to flag them correctly.
As for the stable version, this patch remains the minimal implementation.