[FIX] l10n_de_reports: fix DATEV export if there are return lines
With German localisations and reports.
Open a Point of Sale session.
Make a sale for one product. Make a return for a product.
Close the session and validate.
Go to Accounting, Reporting, Audit Reports, General Ledger.
Choose the current period and click on the button "EXPORT DATEV".
It crashes on singleton error because there is more than one tax line.

If there is more than one line, we sum their values.

opw 1946643

closes odoo/enterprise#3795

Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>