[FIX] l10n_mx_edi: avoid overwrite payment method on payments
When a payment is created independent of a bank statement, usually the
payment have the payment method filled. So, the overriding of this
payment method when a conciliation with a bank statement is made must be
avoided.

Example:

- Create a payment with a payment method filled;
- Sign the payment;
- Get as a PAC status: 'retry' (Maybe because a connection problem);
- Create a bank statement without a payment method to match the payment;
- Conciliate the bank statement with the payment.

Before this commit: The conciliation will overwrite and delete the
payment method. This generates an issue because the payment method is
mandatory to the sign process.

After this commit: The payment method on the payment will stay
unchanged.

closes #4656

closes odoo/enterprise#4714

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>