[FIX] stock_barcode: increment decimals quantity
On the barcode picking interface, when several quantities of a product
are scanned for the same line, the first works as expected but for
succeeding ones the original decimal part is visually dropped when being
incremented with the new quantity.

eg:

- we scan 1.75 kg => 1.75 kg shown
- we scan 0.33 kg => 1.33 kg shown
- we scan 1.25 kg => 2.25 kg shown

The data in Odoo is still correct, if we edit the line or validated it
we would have 3.33kg but visually in the interface the wrong quantity is
shown on the line.

With this changeset we increment float to float and keep 15 precision
digits to avoid insignifiant precision digits:

https://en.wikipedia.org/wiki/Double-precision_floating-point_format

opw-1946049
closes #3811

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>