[FIX] mrp_workorder: use batch_size to decide wether to start of the next workorder
The field 'batch', when set to 'yes', says that the next workorder should start
only after the quantity set in 'Quantity To Process' (batch_size)
has been produced.
Only, that's what it says, not what it does.

We use commit 208b8ebdb0799b78 which created a hook allowing us to override
the behaviour starting the next workorder.

opw 1943674