[IMP] account_asset: Performance improvements
The generation of the deffered revenues took initially 471s.This
improvement permit to reduce the time to 72s (behave Cprofiler).
The issue was related to the fact that all the message post and logs
invalids the cache. It results that several read requests were
repeated multiple times.

closes odoo/odoo#31750

X-port of odoo/odoo@10964a230ca820271926c6b9e45906841e7b0617

Signed-off-by: Denis Ledoux <beledouxdenis@users.noreply.github.com>