[FIX] stock_barcode_quality: Quality checks from barcode views.
It's not possible to validate a picking without doing all the quality checks before, which was not possible in the barcode views.
The cleanest way to fix this is to make a bridge module `stock_barcode_quality` which will auto-install when `stock_barcode` and `quality_control` modules are installed.

To trigger the quality checks wizard, this module will inject 2 buttons in the `stock_barcode` client action.
One in the `SettingsWidget` and another one that will replace the validate button until there are no more quality checks left, in order to unblock the process.

TaskID: 1950619

closes odoo/enterprise#3835

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>