[FIX] documents: open preview while editing
To reproduce, simply click on the document preview while renaming
the same document. The preview flashes on the screen and disappears
shortly after, the page is then locked and the user must reload it.

This is caused by an update of the document state which destroys the
document inspector, destroying the preview as a result.

After this commit, the preview is put on hold until the document is
saved.

opw-1943996

closes odoo/enterprise#3847

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>